test(dedupe): consolidate staleness_helpers tests (5 → 4 files)
Merge data_age_and_severity + freshness_label_and_display into test_staleness_helpers.py
Rename remaining files for consistency:
- banners_and_empty_state → banners
- edge_cases → edge
- execution_health → execution

Cluster: f08c1e19e6ef
