+++ tests/_triage/dedupe_candidates.yaml
@@ -2603,7 +2603,9 @@ clusters:
     expected_file_delta: -4
     priority: low
     rationale: Consolidate 5 test files into single parametrized file
-    status: pending
+    status: done
+    owner: claude
+    pr_url: https://github.com/Solders-Girdles/GPT-Trader/pull/322
     added: '2026-01-20'
   0d3d62f31d0c:
     type: similar_names
