+++ tests/unit/gpt_trader/tui/test_staleness_helpers_edge.py
