+++ tests/unit/gpt_trader/tui/test_staleness_helpers_execution.py
